# Optimize guest code