--- a/Cargo.toml
+++ b/Cargo.toml
@@ -8,4 +8,3 @@ opt-level = 3
 [profile.release]
 debug = 1
 lto = true
-