--- a/common/Cargo.toml
+++ b/common/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "common"
-version = "0.1.0" 
+version = "0.1.0"
 edition = "2021"
 
 [dependencies]