--- a/common/src/lib.rs
+++ b/common/src/lib.rs
@@ -1,5 +1,4 @@
 use num_bigint::BigUint;
-use num_traits::identities::Zero;
 use std::str::FromStr;
 // MODULUS is 2048 mod / number of sequencers for skde
 pub const MODULUS: &str = "109108784166676529682340577929498188950239585527883687884827626040722072371127456712391033422811328348170518576414206624244823392702116014678887602655605057984874271545556188865755301275371611259397284800785551682318694176857633188036311000733221068448165870969366710007572931433736793827320953175136545355129";
@@ -13,7 +12,14 @@ pub struct Input {
     pub base: BigUint,
     pub modulus: BigUint,
     pub range: BigUint,
-    pub result: BigUint, // the result is base^exponent under modulus
+}
+
+#[derive(serde::Serialize, serde::Deserialize)]
+pub struct Output {
+    pub base: BigUint,
+    pub modulus: BigUint,
+    pub range: BigUint,
+    pub u: BigUint,
 }
 
 impl Input {
@@ -28,32 +34,15 @@ impl Input {
         let modulus = BigUint::from_str(modulus_str).expect("Invalid number for Modulus");
         let range = BigUint::from_str(range_str).expect("Invalid number for Range");
 
-        let result = if modulus.is_zero() {
-            BigUint::zero()
-        } else {
-            Self::calculate_private_modular_exponentiation(&base, &modulus)
-        };
-
         println!("Initial parameter settings");
         println!("Base: {}", base);
         println!("Modulus: {}", modulus);
         println!("Range: {}", range);
-        println!("Result of base^exponent % modulus: {}", result);
 
         Input {
             base,
             modulus,
             range,
-            result,
-        }
-    }
-
-    pub fn calculate_private_modular_exponentiation(base: &BigUint, modulus: &BigUint) -> BigUint {
-        let exponent = BigUint::from_str(EXPONENT).expect("Invalid number for Exponent");
-        if modulus.is_zero() {
-            BigUint::zero()
-        } else {
-            base.modpow(&exponent, modulus)
         }
     }
 }