--- a/host/Cargo.toml
+++ b/host/Cargo.toml
@@ -8,6 +8,6 @@ methods = { path = "../methods" }
 common = { path = "../common" }
 num-bigint = { version = "0.4", features = ["rand", "serde"] }
 num-traits = "0.2"
-risc0-zkvm = { version = "1.0.5", features = [ "prove",  "std" ] }
+risc0-zkvm = { version = "1.0.5", features = ["prove", "std"] }
 serde = { version = "1", features = ["derive"] }
 bincode = "1.3"