--- a/host/src/main.rs
+++ b/host/src/main.rs
@@ -1,4 +1,4 @@
-use common::Input;
+use common::{Input, Output};
 use methods::{RANGE_PROOF_ELF, RANGE_PROOF_ID};
 use num_bigint::BigUint;
 use risc0_zkvm::{
@@ -7,38 +7,35 @@ use risc0_zkvm::{
 use std::time::Instant;
 
 fn main() {
-    let (base, modulus, range, result) = setup_inputs();
+    let input = setup_inputs();
 
-    let env = setup_env(&base, &modulus, &range, &result);
+    let env = setup_env(&input.base, &input.modulus, &input.range);
 
     // Generate proof and get receipt
     let receipt = generate_proof(env).receipt;
 
-    let u: BigUint = receipt.journal.decode().unwrap();
+    let output: Output = receipt.journal.decode().unwrap();
 
-    println!("u: {}", u);
-
+    println!("u: {}, range: {}", output.u, output.range);
     // Verify the proof
     verify_proof(&receipt);
 }
 
-fn setup_inputs() -> (BigUint, BigUint, BigUint, BigUint) {
+fn setup_inputs() -> Input {
     let input = Input::new_default();
 
-    (input.base, input.modulus, input.range, input.result)
+    Input {
+        base: input.base,
+        modulus: input.modulus,
+        range: input.range,
+    }
 }
 
-fn setup_env<'a>(
-    base: &'a BigUint,
-    modulus: &'a BigUint,
-    range: &'a BigUint,
-    result: &'a BigUint,
-) -> ExecutorEnv<'a> {
+fn setup_env<'a>(base: &'a BigUint, modulus: &'a BigUint, range: &'a BigUint) -> ExecutorEnv<'a> {
     let input = Input {
         base: base.clone(),
         modulus: modulus.clone(),
         range: range.clone(),
-        result: result.clone(),
     };
 
     ExecutorEnv::builder()