--- a/methods/Cargo.toml
+++ b/methods/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [build-dependencies]
-risc0-build = { version = "1.0.5" }
+risc0-build = { version = "2.0.1" }
 
 [package.metadata.risc0]
 methods = ["guest"]