--- a/methods/guest/Cargo.toml
+++ b/methods/guest/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 [workspace]
 
 [dependencies]
-num-bigint = {version = "0.4.4", features = [ "serde" ]}
+num-bigint = { version = "0.4.4", features = ["serde"] }
 num-traits = "0.2.16"
 risc0-zkvm = { version = "1.0.5", default-features = false, features = ['std'] }
 common = { path = "../../common" }