--- a/methods/guest/Cargo.toml
+++ b/methods/guest/Cargo.toml
@@ -8,7 +8,9 @@ edition = "2021"
 [dependencies]
 num-bigint = { version = "0.4.4", features = ["serde"] }
 num-traits = "0.2.16"
-risc0-zkvm = { version = "1.0.5", default-features = false, features = ['std'] }
+risc0-zkvm = { version = "2.0.1", default-features = false, features = [
+    'std',
+] }
 common = { path = "../../common" }
 bincode = "1.3"
 serde = { version = "1", features = ["derive"] }