This PR closes issue #3 

# Changes
- Remove Unnecessary variables and logics in the guset code
- Apply succinct receipt(recursive zk stark proof + zk snark)
- Benchmark based on the type of receipt.


  | Composite Receipt | Fast Receipt | Succinct Receipt
-- | -- | -- | --
proving method | zk stark | zk snark | zk stark + snark
Hash function | Poseidon2 | sha256 | Poseidon2
Proving time | 126.383148166s | **114.381178625s** | 154.6101965s
Verifying time | 123.834666ms | 18.826084ms | **10.545125ms**