@woojucy 2025-04-04
In the host side code, the `input` is being set, and this `input` is then accessed from the guest side code. I believe that during the verification process, the input should be reflected in the `receipt` object. It would be helpful if this part could be clarified and properly reflected in the code.