@energyGiver 2025-04-07
> In the host side code, the `input` is being set, and this `input` is then accessed from the guest side code. I believe that during the verification process, the input should be reflected in the `receipt` object. It would be helpful if this part could be clarified and properly reflected in the code.

@woojucy Thank you for your comment. According to https://docs.rs/risc0-zkvm/2.0.1/risc0_zkvm/guest/index.html, public input is set using env::commit, and other inputs are private.

I modified exponent as an input, and changed the public output like this:
env::commit(&(base, modulus, range, calculation));